+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Procedure when bumping the version number:
 - Config file
 - `ephemeral` config option
 - `data_dir` config option
+- `euph.rooms.<name>.autojoin` config option
 - `euph.rooms.<name>.username` config option
 - `euph.rooms.<name>.force_username` config option
 - `euph.rooms.<name>.password` config option